End of day 5
Omg what a day, did so much but accomplished so little, wasted a lot of time on the entity bug that took me 3 hours to hunt down, also the god damn ore tiles where a pain in the ass to know how to program. Ended up just using an oreID and putting it in the chunkdata, and then when we are destroying a tile, check if that tile has an oreID in the chunkdata and go from there. Didn't actually see if breaking works but I think it should. Tomorrow will do some biome stuff, lucy made nice looking tiles that I put in. Starting to look a lot better already!