Interior
OMG I can't believe that took so long. Multiplayer is HARD man why is it so complicated. It is not really if you think about it but there is this extra layer to think about, and it just makes it harder to debug and know what is actually going on. And also I have no clue how fishnet actually works and I'm struggling out here but we can go into the interior now and everything (not entities because they don't have world as parent yet) get hidden