Better worldgen
Removed the cellular automata shit because it didn't even do anything, just using noise now and it looks very fine, it is fine like this, look great. We can make each biome have a different noise and shit and that will be perfectly fine. Also surface uses actual noise as well now. Good night

Tomorrow I don't know what I'll do, we need an inventory system because we'll need to place down blocks, this means we'll need a hot bar as well and a max capacity omg this will be some work, but thats for tomorrow!