End of day 16
So, I fucked around with Odin inspector for like 2 hours, which I don't know if it was worth my time, atleast I know how to fucking divide it into two groups, what a nightmare, it looks somewhat cool but idk. We'll see if it will actually be useful, it probably will be a little bit nicer to add new scriptable objects now that we got this. Anyway, tomorrow I'll fix the entity spawning, probably add more entity stuff, I want to decorate the so it looks better and dynamic. Lucy made a cool ship wreck which I have no clue how I will spawn, but there will probably be a way. Probably need to spawn it just once, or have a set location it spawns at.. IDK that is for tomorrow, good night