End of day 20
Okay, I'm done fucking with it. It seems like the bottleneck is not really the entity spawning anyway, there is some kind of architecture for it now, incase we want to make it faster later. But that is not our problem right now. We have to continue with the actual content, and optimize later. No need to chase random things right now, you just need to keep going and actually make the game. I'm done fucking around, just got to keep going.

I do like the async nature of it though, we can just request, and then we got to wait until all the data is there, and then render it, we don't halt the entire main thread until we are done, which is ALOT smoother than how it was like before