Interior sub fixes
Mainly player animations and background. I need to do some kind of entity rework now because I've realized it is two systems that can kind of just be one. Basically I want interior objects to be treated as Entities. And with this modification, WorldVisibilityManager should idealy also use entities, well, it is kind of already doing that, so I don't think we need to change much there. All we have to do is just to register the interior objects as entities within the entity manager