Start of Fish
Two months in and we are finally adding fish wop wop. Did some better visual things also some very basic placeholder visual for the waste removal, it does NOT work with multiplayer, which is such a pain in the ass and so much more work why are we not just doing a singleplayer game!? Oh well this will all be worth it in the end I hope, trying to speed run the trailer tbh!