World gen
Omg what a nightmare to get this working but now we can actually use HLSL and put the values in CODE, instead of into this horrible shader graph box. Also we could now also just loop through all the biomes and not have a huge shader graph. Anyway good night!