More bug fixing
This visual tool stuff is killing me man what the fuck why is it SO buggy I can't believe it. Need to still have the lazer fixed, and I'm still not sure if the actual toolController is registering the remote and local clients properly.  Oh yeah also updated hot reload and fishnet