Entity Spawn Event
Why was this so hard to get right? I still don't know if I should store the data in the EntitySpecific data, or just within the class, it is so confusing, I don't know which one I should do. I don't think it really matters, I think I might be overthinking it, I'm not sure what to do. I'll just have to pick one and stop overthinking