Upgrade system done
Well, there is probably something wrong it with, but right now it seems to be working fine. Needs a lot more content work, but the base is there. Now every tree needs to be designed and linked to the data. But it feels like that is the only good way to do it like this. It definitely makes it a lot more flexible