Start of upgrade visuals
Data is there, data is getting read and displayed, now all that is left is hooking up the buttons and contributing to the upgrade, and making that sync with everything, then also storing and going to the next upgrade, then it should be all done! Making good progress! Now time for some SkyFactory4