Goodnight
Next up is to get the background working, I feel like this is even tricker, its alike another layer ontop of the world shader, but I feel like its quite easy because we just copy what we do in the world shader and slap on some effects, shouldn't be so difficult right? Anyway goodnight!