World gen cleanup
Removed some old stuff. Created a new manager which handles settings and will do more later. The problem was that the world gen settings where only created locally for the server and if you joined later the background shader wouldn't work because the values are pulled from the scriptable object but that only happened on the server so now we just save it as the scriptable object and pull the ID from the server, later we probably want to make this as a string or some kind of struct with a million different parameters which will need to be sent over the network just because the background shader relies on all those parameters to work properly