Fixed background edges
Using screen space DDX and DDY to check the derivative for the SDF so we can finally have uniform looking edges. Problem now though is that if we zoom out it will keep the same edges thickness no matter how far away you are from the background. But we could probably hack it and change the thickness depending on that distance or something