UI shows temp buff
But now I want the UI to show the actual ability all the time because you want to know when you can activate it. But this is giving me a problem because how do I do that?? The actual buff we get from the ability is something we don't want to display. We just want to show that we have it activated. We need a new manager or something that holds actual ability data and the things we have unlocked onto our character, this is where we then also later would add passive abilities. I think we need to rename the things. We have temporary and permanent BUFFS, this is what we've implemented now, but now we need temporary and permanent ABILITIES. All the tools are technically permanent abilities, they are only manual ones you have to click and aim with to use, which is really weird to think of them like that, but they are, and within those permanent abilities, we have a temporary ability, which we can activate. If we implement it like this, we can then add modifiers to these abilities which make them stronger. Like reducing the time it takes for the ability to fire, or making the passive ability stronger, etc.  If we implement a system like this, adding other permanent abilities will be easy, they will just be permanent but also passive, meaning you don't have to do anything, they'll just destroy blocks for you. Yeah lets do it!