Reworked Tools
Oh my again we did the tools AGAIN. It's not working with multiplayer now. But that is not the goal anyway. And only the actual lazer works with the new system. But the new system is so much easier to work with now and will work a lot better with what we actually want the game to be like so that is good, shoudn't take too long to also make the other tools work. But also adding more ailities will be easier onw aswell