Buff upgrade
Yes, we can now upgrade the buff of the brimstone lazer, we'll have to see if this will work with other buff types, and not just the lazer, but it is a good start.

Also the popup shows the buff strength, yes, that is how it will work now.

Also buffs can sit on ability instances and they will be used within the calculation when we add the buff. It's very cool